requirements.txt
influxdb
pyserial